Task2/book_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
python-jose
python-multipart
passlib